ADD: Richardson–Lucy deconvolve CHANGE: use int for color_type (TODO: support several types in Plane/ImageEx)

+++ src/color.hpp
@@ -22,7 +22,7 @@
 #include <QColor>
 #include <cmath>
 
-typedef unsigned short color_type;
+typedef int color_type;
 
 struct color{
 	color_type r;
